requirements.txt
pytest==8.3.4
playwright==1.48.0
pytest-playwright==0.5.0
python-dotenv==1.2.1
allure-pytest==2.14.0
